# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Booking System API for Kosovo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
        - name: barberId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    delete:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /appointments/{id}/confirm-otp:
    post:
      operationId: confirmAppointmentOtp
      tags: [appointments]
      summary: Confirm appointment with OTP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpConfirmInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/resend-otp:
    post:
      operationId: resendAppointmentOtp
      tags: [appointments]
      summary: Resend OTP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OTP sent
  /available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a barber
      parameters:
        - name: shopId
          in: query
          required: true
          schema:
            type: integer
        - name: barberId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlots"

  # PRODUCTS
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [user, owner]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, owner, user]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Barbershop:
      type: object
      required: [id, ownerId, name, city, address, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        city:
          type: string
        address:
          type: string
        description:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, rejected, suspended]
        subdomain:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Barber:
      type: object
      required: [id, shopId, name, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        specialties:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Service:
      type: object
      required: [id, shopId, name, price, durationMinutes, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Appointment:
      type: object
      required: [id, userId, shopId, barberId, serviceId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        shopId:
          type: integer
        barberId:
          type: integer
        serviceId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending_otp, confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        totalPrice:
          type: ["number", "null"]
        user:
          $ref: "#/components/schemas/User"
        barbershop:
          $ref: "#/components/schemas/Barbershop"
        barber:
          $ref: "#/components/schemas/Barber"
        service:
          $ref: "#/components/schemas/Service"
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [shopId, barberId, serviceId, scheduledAt]
      properties:
        shopId:
          type: integer
        barberId:
          type: integer
        serviceId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
    AppointmentList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
    OtpConfirmInput:
      type: object
      required: [otpCode]
      properties:
        otpCode:
          type: string
          minLength: 6
          maxLength: 6
    AvailableSlots:
      type: object
      required: [date, slots]
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            type: string

    # PRODUCT
...