# lib/api-zod/src/generated/types/barber.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Booking System API for Kosovo
 * OpenAPI spec version: 0.1.0
 */

export interface Barber {
  id: number;
  shopId: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  specialties?: string | null;
  /** @nullable */
  rating?: number | null;
  isActive?: boolean;
  createdAt: Date;
}