# lib/api-zod/src/generated/types/barbershop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Booking System API for Kosovo
 * OpenAPI spec version: 0.1.0
 */
import type { BarbershopStatus } from './barbershopStatus';

export interface Barbershop {
  id: number;
  ownerId: number;
  name: string;
  city: string;
  address: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  status: BarbershopStatus;
  /** @nullable */
  subdomain?: string | null;
  /** @nullable */
  rating?: number | null;
  totalReviews?: number;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  openTime?: string | null;
  /** @nullable */
  closeTime?: string | null;
  createdAt: Date;
}