# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Booking System API for Kosovo
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  shopId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  durationMinutes: number;
  createdAt: Date;
}